@@ -37,6 +37,6 @@
     "prettier": "^2.2.1",
     "semantic-release": "^19.0.2",
     "ts-jest": "^28.0.0",
-    "typescript": "^4.4.3"
+    "typescript": "^4.7.4"
   }
 }
