@@ -1,11 +1,13 @@
-export function cx(): string {
+type Argument = string | boolean | number;
+
+export function cx(...args: Argument[]): string {
   let str = "",
     i = 0,
     arg: unknown,
     val: string | number;
 
-  while (i < arguments.length) {
-    if ((arg = arguments[i++])) {
+  while (i < args.length) {
+    if ((arg = args[i++])) {
       if (
         (val = typeof arg === "string" || typeof arg === "number" ? arg : "")
       ) {
